Renames `ModeService` to `LanguageService`
