+++ src/vs/editor/standalone/test/browser/standaloneLanguages.test.ts
@@ -11,7 +11,7 @@ import { Token } from 'vs/editor/common/core/token';
 import { IState, LanguageId, MetadataConsts } from 'vs/editor/common/modes';
 import { ModesRegistry } from 'vs/editor/common/modes/modesRegistry';
 import { TokenTheme } from 'vs/editor/common/modes/supports/tokenization';
-import { ModeServiceImpl } from 'vs/editor/common/services/modeServiceImpl';
+import { LanguageService } from 'vs/editor/common/services/languageServiceImpl';
 import { ILineTokens, IToken, TokenizationSupport2Adapter, TokensProvider } from 'vs/editor/standalone/browser/standaloneLanguages';
 import { IStandaloneTheme, IStandaloneThemeData, IStandaloneThemeService } from 'vs/editor/standalone/common/standaloneThemeService';
 import { ColorIdentifier } from 'vs/platform/theme/common/colorRegistry';
@@ -113,7 +113,7 @@ suite('TokenizationSupport2Adapter', () => {
 		}
 
 		const disposables = new DisposableStore();
-		const languageService = disposables.add(new ModeServiceImpl());
+		const languageService = disposables.add(new LanguageService());
 		disposables.add(ModesRegistry.registerLanguage({ id: languageId }));
 		const adapter = new TokenizationSupport2Adapter(
 			languageId,
