+++ src/vs/workbench/contrib/notebook/test/browser/testNotebookEditor.ts
@@ -18,7 +18,7 @@ import { ILanguageConfigurationService } from 'vs/editor/common/modes/languageCo
 import { IModelService } from 'vs/editor/common/services/modelService';
 import { ModelServiceImpl } from 'vs/editor/common/services/modelServiceImpl';
 import { ILanguageService } from 'vs/editor/common/services/languageService';
-import { ModeServiceImpl } from 'vs/editor/common/services/modeServiceImpl';
+import { LanguageService } from 'vs/editor/common/services/languageServiceImpl';
 import { ITextModelService } from 'vs/editor/common/services/resolverService';
 import { TestLanguageConfigurationService } from 'vs/editor/test/common/modes/testLanguageConfigurationService';
 import { BrowserClipboardService } from 'vs/platform/clipboard/browser/clipboardService';
@@ -158,7 +158,7 @@ export class NotebookEditorTestModel extends EditorModel implements INotebookEdi
 
 export function setupInstantiationService(disposables = new DisposableStore()) {
 	const instantiationService = new TestInstantiationService();
-	instantiationService.stub(ILanguageService, disposables.add(new ModeServiceImpl()));
+	instantiationService.stub(ILanguageService, disposables.add(new LanguageService()));
 	instantiationService.stub(IUndoRedoService, instantiationService.createInstance(UndoRedoService));
 	instantiationService.stub(IConfigurationService, new TestConfigurationService());
 	instantiationService.stub(IThemeService, new TestThemeService());
