+++ src/vs/workbench/services/keybinding/test/browser/keybindingEditing.test.ts
@@ -9,7 +9,7 @@ import { KeyCode } from 'vs/base/common/keyCodes';
 import { ChordKeybinding, SimpleKeybinding } from 'vs/base/common/keybindings';
 import { OS } from 'vs/base/common/platform';
 import { ILanguageService } from 'vs/editor/common/services/languageService';
-import { ModeServiceImpl } from 'vs/editor/common/services/modeServiceImpl';
+import { LanguageService } from 'vs/editor/common/services/languageServiceImpl';
 import { IModelService } from 'vs/editor/common/services/modelService';
 import { ModelServiceImpl } from 'vs/editor/common/services/modelServiceImpl';
 import { ITextModelService } from 'vs/editor/common/services/resolverService';
@@ -105,7 +105,7 @@ suite('KeybindingsEditing', () => {
 		instantiationService.stub(IEditorService, new TestEditorService());
 		instantiationService.stub(IWorkingCopyService, disposables.add(new WorkingCopyService()));
 		instantiationService.stub(ITelemetryService, NullTelemetryService);
-		instantiationService.stub(ILanguageService, ModeServiceImpl);
+		instantiationService.stub(ILanguageService, LanguageService);
 		instantiationService.stub(ILogService, new NullLogService());
 		instantiationService.stub(ILabelService, disposables.add(instantiationService.createInstance(LabelService)));
 		instantiationService.stub(IFilesConfigurationService, disposables.add(instantiationService.createInstance(FilesConfigurationService)));
