+++ src/vs/workbench/test/electron-browser/testing.ts
@@ -5,7 +5,7 @@
 
 import * as assert from 'assert';
 import { LanguagesRegistry } from 'vs/editor/common/services/languagesRegistry';
-// import { ModeServiceImpl } from 'vs/editor/common/services/modeServiceImpl';
+// import { LanguageService } from 'vs/editor/common/services/languageServiceImpl';
 
 /**
  * This function is called before test running and also again at the end of test running
@@ -14,6 +14,6 @@ import { LanguagesRegistry } from 'vs/editor/common/services/languagesRegistry';
 export function assertCleanState(): void {
 	// If this test fails, it is a clear indication that
 	// your test or suite is leaking services (e.g. via leaking text models)
-	// assert.strictEqual(ModeServiceImpl.instanceCount, 0, 'No leaking ILanguageService');
+	// assert.strictEqual(LanguageService.instanceCount, 0, 'No leaking ILanguageService');
 	assert.strictEqual(LanguagesRegistry.instanceCount, 0, 'No leaking LanguagesRegistry');
 }
