+++ src/vs/workbench/test/electron-browser/textsearch.perf.integrationTest.ts
@@ -11,7 +11,7 @@ import * as path from 'vs/base/common/path';
 import { URI } from 'vs/base/common/uri';
 import { IModelService } from 'vs/editor/common/services/modelService';
 import { ModelServiceImpl } from 'vs/editor/common/services/modelServiceImpl';
-import { ModeServiceImpl } from 'vs/editor/common/services/modeServiceImpl';
+import { LanguageService } from 'vs/editor/common/services/languageServiceImpl';
 import { ITextResourcePropertiesService } from 'vs/editor/common/services/textResourceConfigurationService';
 import { TestLanguageConfigurationService } from 'vs/editor/test/common/modes/testLanguageConfigurationService';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
@@ -86,7 +86,7 @@ suite.skip('TextSearch performance (integration)', () => {
 						new TestThemeService(),
 						logService,
 						undoRedoService,
-						new ModeServiceImpl(),
+						new LanguageService(),
 						new TestLanguageConfigurationService()
 					),
 				],
